requirements.txt
biom-format>=2.1.10
pandas==1.5.3
matplotlib>=3.5.0
seaborn>=0.11.1
scipy==1.10.0
numpy>=1.21.0
scikit-learn>=0.24.2
Biopython==1.83
scikit-bio==0.5.8
